+++ freemocap/data_layer/export_data/blender_stuff/blender_bpy_export_scripts/cgtinker_blendarmocap_load.py
