+++ freemocap/data_layer/recording_models/recording_info_model.py
