+++ freemocap/gui/qt/utilities/update_most_recent_recording_toml.py
@@ -2,7 +2,7 @@ import logging
 
 import toml
 
-from freemocap.recording_models.recording_info_model import RecordingInfoModel
+from freemocap.data_layer.recording_models.recording_info_model import RecordingInfoModel
 from freemocap.system.paths_and_filenames.path_getters import get_most_recent_recording_toml_path
 
 logger = logging.getLogger(__name__)
