+++ freemocap/gui/qt/widgets/control_panel/process_mocap_data_panel/parameter_groups/create_parameter_groups.py
@@ -1,6 +1,6 @@
 from pyqtgraph.parametertree import Parameter
 
-from freemocap.recording_models.post_processing_parameter_models import MediapipeParametersModel, \
+from freemocap.data_layer.recording_models.post_processing_parameter_models import MediapipeParametersModel, \
     PostProcessingParameterModel, AniposeTriangulate3DParametersModel, PostProcessingParametersModel, \
     ButterworthFilterParametersModel
