+++ freemocap/utilities/create_nested_dict_from_pydantic.py
@@ -29,6 +29,6 @@ def create_nested_dict(pydantic_model):
 if __name__ == "__main__":
     from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.mediapipe_skeleton_names_and_connections import \
         mediapipe_skeleton_schema
-    from freemocap.data_saver.data_models import SkeletonSchema
+    from freemocap.data_layer.data_saver import SkeletonSchema
     skeleton_schema = SkeletonSchema(schema_dict=mediapipe_skeleton_schema)
     print(create_nested_dict(skeleton_schema))
